authentication system updated

+++ frontend/src/components/Form.jsx
@@ -1,16 +1,17 @@
 import { useState } from "react";
 import api from "../api";
-import { useNavigate } from "react-router-dom";
-import { ACCESS_TOKEN, REFRESH_TOKEN ,USER_NAME} from "../constants";
+import { Link, useNavigate } from "react-router-dom";
+import { ACCESS_TOKEN, REFRESH_TOKEN, USER_NAME } from "../constants";
 import "../styles/Form.css"
 import LoadingIndicator from "./LoaginIndicator";
 
 function Form({ route, method }) {
-    
-   
-    
+
+
+
     const [username, setUsername] = useState("");
     const [password, setPassword] = useState("");
+    const [confirmPassword, setConfirmPassword] = useState("");
     const [loading, setLoading] = useState(false);
     const navigate = useNavigate();
 
@@ -19,20 +20,26 @@ function Form({ route, method }) {
 
 
     const handleSubmit = async (e) => {
-        console.log(route,method);
+        console.log(route, method);
         setLoading(true);
         e.preventDefault();
 
         try {
+
+            if(method=="register" && password !==confirmPassword){
+                alert("Password does not match");
+                return;
+            }
             const res = await api.post(route, { username, password });
-            
+
             if (method === "login") {
-               
+
                 localStorage.setItem(ACCESS_TOKEN, res.data.access);
                 localStorage.setItem(REFRESH_TOKEN, res.data.refresh);
-                localStorage.setItem(USER_NAME,username);
+                localStorage.setItem(USER_NAME, username);
                 navigate("/")
             } else {
+
                 localStorage.clear();
                 navigate("/login")
             }
@@ -44,7 +51,7 @@ function Form({ route, method }) {
     };
 
     return (
-        <form  className="form-container" onSubmit={(e)=>handleSubmit(e)}>
+        <form className="form-container" onSubmit={(e) => handleSubmit(e)}>
             <h1>{name}</h1>
             <input
                 className="form-input"
@@ -60,10 +67,25 @@ function Form({ route, method }) {
                 onChange={(e) => setPassword(e.target.value)}
                 placeholder="Password"
             />
+            {method == "register" ?
+                <input
+                    className="form-input"
+                    type="password"
+                    value={confirmPassword}
+                    onChange={(e) => setConfirmPassword(e.target.value)}
+                    placeholder="Confirm Password"
+                />
+                : null
+            }
+
+
+
             {loading && <LoadingIndicator />}
             <button className="form-button" type="submit">
                 {name}
             </button>
+
+            <Link to={method == "register" ? '/login' : '/register'}>{method == 'register' ? "Login" : "Register"}</Link>
         </form>
     );
 }
